+++ test/sys_tests/test_sys_single_jump.py
@@ -16,7 +16,9 @@ def main():
                    si_instance_type='t2.micro',
                    vpc=vpc,
                    subnet=subnet,
-                   template=template)
+                   template=template,
+                   instance_dependencies=vpc.title,
+                   hosted_zone_name=None)
 
     template.add_resource(vpc)
     template.add_resource(subnet)