+++ test/unit_tests/test_yaml.py
@@ -8,7 +8,7 @@ from nose.tools import *
 from cerberus import ValidationError
 from amazonia.classes.yaml import Yaml, InsecureVariableError
 
-default_data = None
+default_data = schema = None
 
 
 def setup_resources():
@@ -59,7 +59,7 @@ def test_complete_valid_values():
 
     # Assert correct values
     assert_equals(stack_input['stack_title'], 'testStack')
-    assert_equals(stack_input['code_deploy_service_role'], 'arn:aws:iam::1234567890124:role/CodeDeployServiceRole')
+    assert_equals(stack_input['code_deploy_service_role'], 'arn:aws:iam::1234567890123:role/CodeDeployServiceRole')
     assert_equals(stack_input['keypair'], 'key')
     assert_list_equal(stack_input['availability_zones'], ['ap-southeast-2a', 'ap-southeast-2b', 'ap-southeast-2c'])
     assert_equals(stack_input['vpc_cidr'], '10.0.0.0/16')
@@ -68,8 +68,9 @@ def test_complete_valid_values():
     assert_equals(stack_input['jump_instance_type'], 't2.micro')
     assert_equals(stack_input['nat_image_id'], 'ami-53371f30')
     assert_equals(stack_input['nat_instance_type'], 't2.micro')
+    assert_equals(stack_input['stack_hosted_zone_name'], 'test.org.')
     assert_equals(type(stack_input['home_cidrs']), list)
-    assert_equals(len(stack_input['home_cidrs']), 2)
+    assert_equals(len(stack_input['home_cidrs']), 1)
     assert_equals(type(stack_input['autoscaling_units']), list)
     assert_equals(len(stack_input['autoscaling_units']), 2)
     assert_equals(type(stack_input['database_units']), list)
@@ -89,7 +90,6 @@ def test_complete_valid_values():
     assert_equals(len(expected_autoscaling_unit_set.difference(autoscaling_unit_input_set)), 0)
 
     assert_equals(autoscaling_unit_input['unit_title'], 'app1')
-    assert_equals(autoscaling_unit_input['hosted_zone_name'], '.test.lan')
     assert_equals(autoscaling_unit_input['image_id'], 'ami-dc361ebf')
     assert_equals(autoscaling_unit_input['instance_type'], 't2.micro')
     assert_equals(autoscaling_unit_input['path2ping'], '/index.html')
@@ -101,10 +101,17 @@ def test_complete_valid_values():
     assert_equals(autoscaling_unit_input['health_check_grace_period'], '300')
     assert_equals(autoscaling_unit_input['iam_instance_profile_arn'], 'arn:aws:iam::1234567890124:role/InstanceProfile')
     assert_equals(autoscaling_unit_input['sns_topic_arn'], 'sns_topic_arn')
-    assert_equals(autoscaling_unit_input['sns_notification_types'], ['sns_notification_types'])
+    assert_equals(autoscaling_unit_input['sns_notification_types'], ['autoscaling:EC2_INSTANCE_LAUNCH',
+                                                                     'autoscaling:EC2_INSTANCE_LAUNCH_ERROR',
+                                                                     'autoscaling:EC2_INSTANCE_TERMINATE',
+                                                                     'autoscaling:EC2_INSTANCE_TERMINATE_ERROR'])
     assert_equals(autoscaling_unit_input['elb_log_bucket'], 'elb_log_bucket')
     assert_equals(autoscaling_unit_input['health_check_type'], 'ELB')
     assert_list_equal(autoscaling_unit_input['dependencies'], ['app2', 'db1'])
+    assert_equals(autoscaling_unit_input['unit_hosted_zone_name'], 'test.org.')
+
+    # Assert that unit is picking up the stack_hosted_zone_name if unit_hosted_zone_name isn't provided
+    assert_equals(stack_input['autoscaling_units'][0]['unit_hosted_zone_name'], stack_input['stack_hosted_zone_name'])
 
     database_unit_input = stack_input['database_units'][0]
 
@@ -123,6 +130,7 @@ def test_complete_valid_values():
     assert_equals(database_unit_input['db_instance_type'], 'db.m1.small')
     assert_equals(database_unit_input['db_engine'], 'postgres')
     assert_equals(database_unit_input['db_port'], '5432')
+    assert_equals(database_unit_input['db_name'], 'myDb')
 
 
 @with_setup(setup_resources())
@@ -183,6 +191,7 @@ def test_insecure_variables_yaml():
                                                   'default_data': default_data,
                                                   'schema': schema})
 
+
 @with_setup(setup_resources())
 def test_invalid_min_max_asg():
     """
@@ -191,7 +200,7 @@ def test_invalid_min_max_asg():
     global default_data
     invalid_min_max_stack_data = open_yaml_file('test_yaml_invalid_min_max_asg.yaml')
 
-    assert_raises(ValidationError, Yaml, **{'user_stack_data':invalid_min_max_stack_data,
+    assert_raises(ValidationError, Yaml, **{'user_stack_data': invalid_min_max_stack_data,
                                             'default_data': default_data,
                                             'schema': schema})