+++ CHANGELOG.md
@@ -2,7 +2,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.14] - 04/06/2016
+## [1.0.17] - 08/07/2016
+- Adding s3 Class, Unit tests, System tests
+- Adding Cloudtrail Class, Unit tests, Systems tests
+- Updated yaml to be more consistant
+- Updated yaml and tests to be more encompassing
+- Removed test inner and intra test stubs (will return soon)
+- Updated test_sys_stack_create to allow passing of cloud formation parameters
+
+## [1.0.16] - 07/07/2016
+- Updated Hosted_zone_name to be passed at stack level or unit level in yaml as stack_hosted_zone_name or unit_hosted_zone_name
+- Updated single_instance to create an elastic IP and a route 53 record set for jump hosts if stack_hosted_zone_name is provided
+- Updated 'DependsOn' options for single instance to be passed in rather than assigned after creation.
+- Added 'db_name' yaml value to match DBName value in RDS instances.
+- Added 'hdd_size' yaml value to change the size of Autoscaling unit instances.
+
+## [1.0.14] - 04/07/2016
 - Updated amazonia/test/sys_tests/test_sys_stack_create.py to include cases for CREATE_FAILED, DELETE_FAILED, ROLLBACK_FAILED for create_and_delete_stack()
 
 ## [1.0.9] - 30/06/2016